move the help message code to an observer
