+++ plugins/nf-schema/src/resources/META-INF/extensions.idx
@@ -1,1 +1,2 @@
-nextflow.validation.SchemaValidator
+nextflow.validation.ValidationExtension
+nextflow.validation.ValidationObserverFactory
